# server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/employees": {
      "post": {
        "operationId": "EmployeesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "従業員作成",
        "tags": [
          "employees"
        ]
      },
      "get": {
        "operationId": "EmployeesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "従業員一覧",
        "tags": [
          "employees"
        ]
      }
    },
    "/api/v1/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "従業員詳細",
        "tags": [
          "employees"
        ]
      },
      "patch": {
        "operationId": "EmployeesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "従業員更新",
        "tags": [
          "employees"
        ]
      }
    },
...
    "/api/v1/leave-requests": {
      "post": {
        "operationId": "LeaveRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeaveRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "休暇申請作成",
        "tags": [
          "leave-requests"
        ]
      },
      "get": {
        "operationId": "LeaveRequestsController_findMine",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "自身の休暇申請一覧",
        "tags": [
          "leave-requests"
        ]
      }
    },
    "/api/v1/admin/leave-requests": {
      "get": {
        "operationId": "LeaveRequestsController_findAllAdmin",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequestAdminListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "テナント内申請一覧（tenant_admin）",
        "tags": [
          "leave-requests"
        ]
      }
    },
    "/api/v1/admin/leave-requests/{id}/approve": {
      "post": {
        "operationId": "LeaveRequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "承認",
        "tags": [
          "leave-requests"
        ]
      }
    },
    "/api/v1/admin/leave-requests/{id}/reject": {
      "post": {
        "operationId": "LeaveRequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionMessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "差し戻し",
        "tags": [
          "leave-requests"
        ]
      }
    },
    "/api/v1/admin/closing": {
      "post": {
        "operationId": "ClosingController_close",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClosingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosingRecordResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "締め実行",
        "tags": [
          "closing"
        ]
      },
      "get": {
        "operationId": "ClosingController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosingRecordListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "締め履歴一覧",
        "tags": [
          "closing"
        ]
      }
    },
    "/api/v1/admin/closing/{id}/reopen": {
      "post": {
        "operationId": "ClosingController_reopen",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosingRecordResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "締め再開",
        "tags": [
          "closing"
        ]
      }
    },
...
  },
  "info": {
    "title": "勤怠管理システム API",
    "description": "マルチテナント勤怠管理システムのREST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "employeeNumber": {
            "type": "string",
            "example": "EMP001"
          },
          "name": {
            "type": "string",
            "example": "山田太郎"
          },
          "joinedAt": {
            "type": "string",
            "example": "2024-04-01"
          },
          "userId": {
            "type": "string",
            "description": "紐付けるユーザーID"
          },
          "departmentId": {
            "type": "string",
            "description": "部署ID"
          },
          "workPatternId": {
            "type": "string",
            "description": "勤務形態ID"
          }
        },
        "required": [
          "employeeNumber",
          "name",
          "joinedAt"
        ]
      },
      "DepartmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name"
        ]
      },
      "WorkPatternResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "breakMinutes": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "tenantId",
          "name",
          "startTime",
          "endTime",
          "breakMinutes"
        ]
      },
      "EmployeeUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "system_admin",
              "tenant_admin",
              "tenant_user"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "EmployeeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "userId": {
            "type": "object",
            "nullable": true
          },
          "employeeNumber": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "departmentId": {
            "type": "object",
            "nullable": true
          },
          "workPatternId": {
            "type": "object",
            "nullable": true
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "department": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/DepartmentResponseDto"
              }
            ]
          },
          "workPattern": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkPatternResponseDto"
              }
            ]
          },
          "user": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/EmployeeUserResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "tenantId",
          "userId",
          "employeeNumber",
          "name",
          "departmentId",
          "workPatternId",
          "joinedAt",
          "createdAt",
          "department",
          "workPattern",
          "user"
        ]
      },
      "EmployeeListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmployeeResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "workPatternId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        }
      },
...
      "AttendanceEmployeeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "employeeNumber": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "employeeNumber"
        ]
      },
...
      "ActionMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Approved"
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateLeaveRequestDto": {
        "type": "object",
        "properties": {
          "leaveType": {
            "type": "string",
            "enum": [
              "paid",
              "unpaid",
              "sick",
              "other"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2026-05-01"
          },
          "endDate": {
            "type": "string",
            "example": "2026-05-03"
          },
          "reason": {
            "type": "string",
            "example": "私用のため"
          }
        },
        "required": [
          "leaveType",
          "startDate",
          "endDate"
        ]
      },
      "LeaveRequestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "leaveType": {
            "type": "string",
            "enum": [
              "paid",
              "unpaid",
              "sick",
              "other"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "reviewedBy": {
            "type": "object",
            "nullable": true
          },
          "reviewedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "employeeId",
          "leaveType",
          "startDate",
          "endDate",
          "reason",
          "status",
          "reviewedBy",
          "reviewedAt",
          "createdAt"
        ]
      },
      "LeaveRequestListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaveRequestResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "LeaveRequestAdminItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "employeeId": {
            "type": "string"
          },
          "leaveType": {
            "type": "string",
            "enum": [
              "paid",
              "unpaid",
              "sick",
              "other"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "reviewedBy": {
            "type": "object",
            "nullable": true
          },
          "reviewedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "employee": {
            "$ref": "#/components/schemas/AttendanceEmployeeResponseDto"
          }
        },
        "required": [
          "id",
          "tenantId",
          "employeeId",
          "leaveType",
          "startDate",
          "endDate",
          "reason",
          "status",
          "reviewedBy",
          "reviewedAt",
          "createdAt",
          "employee"
        ]
      },
      "LeaveRequestAdminListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LeaveRequestAdminItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "CreateClosingDto": {
        "type": "object",
        "properties": {
          "yearMonth": {
            "type": "string",
            "example": "2026-04",
            "description": "対象年月（YYYY-MM）"
          }
        },
        "required": [
          "yearMonth"
        ]
      },
      "ClosingRecordResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "yearMonth": {
            "type": "string"
          },
          "closedBy": {
            "type": "string"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "closed",
              "reopened"
            ]
          }
        },
        "required": [
          "id",
          "tenantId",
          "yearMonth",
          "closedBy",
          "closedAt",
          "status"
        ]
      },
      "ClosingCloserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "ClosingRecordListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "yearMonth": {
            "type": "string"
          },
          "closedBy": {
            "type": "string"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "closed",
              "reopened"
            ]
          },
          "closer": {
            "$ref": "#/components/schemas/ClosingCloserResponseDto"
          }
        },
        "required": [
          "id",
          "tenantId",
          "yearMonth",
          "closedBy",
          "closedAt",
          "status",
          "closer"
        ]
      },
      "ClosingRecordListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClosingRecordListItemResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
...
    }
  }
}